Trim fat. Combine app.js and server.js, add Procfile, simplify configuration.
